Removed unnecessary method from ApplicationUserRepo, refactored ApplicationUserService
