+++ src/main/java/com/project/trackfit/user/IApplicationUserService.java
@@ -1,6 +1,4 @@
-package com.project.trackfit.core;
-
-import com.project.trackfit.user.CreateUserRequest;
+package com.project.trackfit.user;
 
 import java.util.UUID;
